Read image resolution from EXIF, apply to intrinsic size See https://github.com/whatwg/html/pull/5574
Currently applies to JPEG only.

Use the EXIF resolution/size information we read from the
JPEG image to determine a density-corrected preferred size,
which overrides the intrinsic size of the image.

We only apply this intrinsic size when an image has both
resolution and size information in its exif, and if the image
is origin-clean (same origin or cross-origin with CORS).

Bug: 1069755
Change-Id: I01a38c72bb46974a0c112175934cd3c2c48607a9